+++ test/test_network.rb
@@ -2,17 +2,17 @@ require 'helper'
 
 describe 'Network' do
 
-  context 'assuming a bad network' do
+  describe 'assuming a bad network' do
 
-    should 'handle no server available' do
+    it 'handle no server available' do
       assert_raises Dalli::RingError, :message => "No server available" do
         dc = Dalli::Client.new 'localhost:19333'
         dc.get 'foo'
       end
     end
 
-    context 'with a fake server' do
-      should 'handle connection reset' do
+    describe 'with a fake server' do
+      it 'handle connection reset' do
         memcached_mock(lambda {|sock| sock.close }) do
           assert_raises Dalli::RingError, :message => "No server available" do
             dc = Dalli::Client.new('localhost:19123')
@@ -21,7 +21,7 @@ describe 'Network' do
         end
       end
 
-      should 'handle malformed response' do
+      it 'handle malformed response' do
         memcached_mock(lambda {|sock| sock.write('123') }) do
           assert_raises Dalli::RingError, :message => "No server available" do
             dc = Dalli::Client.new('localhost:19123')
@@ -30,7 +30,7 @@ describe 'Network' do
         end
       end
 
-      should 'handle connect timeouts' do
+      it 'handle connect timeouts' do
         memcached_mock(lambda {|sock| sleep(0.6); sock.close }, :delayed_start) do
           assert_raises Dalli::RingError, :message => "No server available" do
             dc = Dalli::Client.new('localhost:19123')
@@ -39,7 +39,7 @@ describe 'Network' do
         end
       end
 
-      should 'handle read timeouts' do
+      it 'handle read timeouts' do
         memcached_mock(lambda {|sock| sleep(0.6); sock.write('giraffe') }) do
           assert_raises Dalli::RingError, :message => "No server available" do
             dc = Dalli::Client.new('localhost:19123')